Add Rust implementation
A follow-on patch will regenerate Make-managed files.

Disclaimer: Participation by NIST in the creation of the documentation
of mentioned software is not intended to imply a recommendation or
endorsement by the National Institute of Standards and Technology, nor
is it intended to imply that any specific software is necessarily the
best available for the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>